Allow empty GOPATH
This is a new feature since Go 1.9, which sets the GOPATH to `$HOME/go` if
the environment variable `GOPATH` is not set.

The implementation is also simplified by using the stdlib API
`filepath.SplitList`. As a side-effect this should also be (a bit) faster
as we don't have to consult the environment variables.